requirements.txt
cohere>=5.0.0
pandas
python-dotenv
